Refactor / split classes / comments
