+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/approximation/ApproximationPlan.java
@@ -67,10 +67,24 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 /**
- * The approximation plan, that is substituted during logical plan optimization
- * in the rule {@link SubstituteApproximationPlan}.
+ * Generates an approximate query for queries suitable for approximation.
  * <p>
- * See the Javadocs of {@link Approximation} for more details.
+ * The approximate query samples documents before the STATS, and extrapolates
+ * the aggregate functions if needed. This new logical plan is generated by
+ * {@link ApproximationPlan#get}. The substitution of the original query by the
+ * approximate query happens during logical plan optimization, in the rule
+ * {@link SubstituteApproximationPlan}.
+ * <p>
+ * In addition to approximate results, confidence intervals are also computed.
+ * This is done by dividing the sampled rows {@link ApproximationPlan#TRIAL_COUNT}
+ * times into {@link ApproximationPlan#BUCKET_COUNT} buckets, computing the aggregate
+ * functions for each bucket, and using these sampled values to compute
+ * confidence intervals with the bias-corrected and accelerated (BCa) bootstrap
+ * method, see also {@link ConfidenceInterval}.
+ * <p>
+ * The initial approximation plan contains a placeholder for the sample probability,
+ * which is determined during subplan execution, see the Javadocs of
+ * {@link Approximation} for more details.
  */
 public class ApproximationPlan {
 
@@ -440,7 +454,7 @@ public class ApproximationPlan {
             }
             originalAggregates.add(agg);
 
-            if (fieldBuckets != null && Approximation.SUPPORTED_SINGLE_VALUED_AGGS.contains(aggFn.getClass())) {
+            if (fieldBuckets != null && ApproximationVerifier.SUPPORTED_SINGLE_VALUED_AGGS.contains(aggFn.getClass())) {
                 // For the supported single-valued aggregations, add buckets with sampled
                 // values, that will be used to compute a confidence interval.
                 // For multivalued aggregations, confidence intervals do not make sense.
@@ -699,13 +713,17 @@ public class ApproximationPlan {
     }
 
     /**
-     * (...)
+     * For FORK, if branches output fields with buckets, these buckets must be
+     * added to the (merged) fork as well. Furthermore, if other fork branches
+     * don't contain the buckets, they must be added in it too with null value.
      */
     private static LogicalPlan forkIncludingBuckets(Fork fork, Map<NameId, List<Attribute>> fieldBuckets) {
         if (fieldBuckets == null) {
             return fork;
         }
 
+        // Check whether the fork output fields have bucket in any branch.
+        // If so, add them to the fork output as well.
         List<Attribute> output = null;
         for (Attribute attribute : fork.output()) {
             children: for (LogicalPlan child : fork.children()) {
@@ -726,10 +744,13 @@ public class ApproximationPlan {
             }
         }
 
+        // If there are no fields with buckets, return the original fork.
         if (output == null) {
             return fork;
         }
 
+        // For each attribute in the fork output, if it's not present in the
+        // output of a branch, add it.
         List<LogicalPlan> children = new ArrayList<>();
         for (LogicalPlan child : fork.children()) {
             Map<String, Attribute> childAttributes = child.output()
@@ -857,7 +878,8 @@ public class ApproximationPlan {
     /**
      * Substitutes the {@link SampleProbabilityPlaceHolder} in the approximation plan
      * by the actual sample probability. If the sample probability is 1.0, the
-     * SampledAggregate is also replaced by a regular Aggregate.
+     * SampledAggregate is also replaced by a regular Aggregate, and the confidence
+     * interval is replaced by a trivial one.
      */
     public static LogicalPlan substituteSampleProbability(LogicalPlan logicalPlan, double sampleProbability) {
         logicalPlan = logicalPlan.transformExpressionsDown(
@@ -897,43 +919,58 @@ public class ApproximationPlan {
 
     /**
      * Substitutes the {@link SampleProbabilityPlaceHolder} in the approximation plan
-     * by the actual sample probability. If the sample probability is 1.0, the
-     * SampledAggregate is also replaced by a regular Aggregate.
+     * in the given fork branch (identified by the branch index) by the actual sample
+     * probability. If the sample probability is 1.0, the SampledAggregate is also
+     * replaced by a regular Aggregate, and the bucket values are set to this
+     * (non-sampled) aggregate value in that fork branch.
      */
-    public static LogicalPlan substituteSampleProbabilityInForkBranch(LogicalPlan logicalPlan, double sampleProbability) {
-        logicalPlan = logicalPlan.transformExpressionsDown(
-            ApproximationPlan.SampleProbabilityPlaceHolder.class,
-            prob -> Literal.fromDouble(Source.EMPTY, sampleProbability)
-        );
-        if (sampleProbability == 1.0) {
-            // TODO: ...
-            logicalPlan = logicalPlan.transformDown(SampledAggregate.class, sampledAggregate -> {
-                List<Alias> fields = new ArrayList<>();
-                int bucketIndex = sampledAggregate.originalAggregates().size();
-                for (int i = 0; i < sampledAggregate.originalAggregates().size(); i++) {
-                    NamedExpression aggOrKey = sampledAggregate.originalAggregates().get(i);
-                    if ((aggOrKey instanceof Alias alias && alias.child() instanceof AggregateFunction) == false) {
-                        continue;
-                    }
-                    for (int j = 0; j < TRIAL_COUNT * BUCKET_COUNT; j++) {
-                        Alias bucket = (Alias) sampledAggregate.aggregates().get(bucketIndex++);
-                        fields.add(bucket.replaceChild(aggOrKey.toAttribute()));
-                    }
-                }
-                return new Eval(
-                    Source.EMPTY,
-                    new Aggregate(
-                        sampledAggregate.source(),
-                        sampledAggregate.child(),
-                        sampledAggregate.groupings(),
-                        sampledAggregate.originalAggregates()
-                    ),
-                    fields
-                );
-            });
-            logicalPlan = new PruneColumns().apply(logicalPlan);
-        }
+    public static LogicalPlan substituteSampleProbabilityInForkBranch(LogicalPlan logicalPlan, double sampleProbability, int branchIndex) {
+        logicalPlan = logicalPlan.transformUp(Fork.class, fork -> {
+            List<LogicalPlan> children = new ArrayList<>(fork.children());
+            assert branchIndex >= 0 && branchIndex < children.size();
 
+            LogicalPlan child = children.get(branchIndex);
+            child = child.transformExpressionsDown(
+                ApproximationPlan.SampleProbabilityPlaceHolder.class,
+                prob -> Literal.fromDouble(Source.EMPTY, sampleProbability)
+            );
+            if (sampleProbability == 1.0) {
+                // When there's no sampling: execute a normal Aggregate, and replace
+                // the bucket values by the non-sampled value. It's incorrect to
+                // replace the confidence interval by a trivial one (like in
+                // `substituteSampleProbability`), because at this stage it's unknown
+                // what happens in other fork branches.
+                child = child.transformDown(SampledAggregate.class, sampledAggregate -> {
+                    List<Alias> fields = new ArrayList<>();
+                    int bucketIndex = sampledAggregate.originalAggregates().size();
+                    for (int i = 0; i < sampledAggregate.originalAggregates().size(); i++) {
+                        NamedExpression aggOrKey = sampledAggregate.originalAggregates().get(i);
+                        if ((aggOrKey instanceof Alias alias && alias.child() instanceof AggregateFunction) == false) {
+                            continue;
+                        }
+                        // Each aggregate is directly followed by its buckets.
+                        for (int j = 0; j < TRIAL_COUNT * BUCKET_COUNT; j++) {
+                            Alias bucket = (Alias) sampledAggregate.aggregates().get(bucketIndex++);
+                            fields.add(bucket.replaceChild(aggOrKey.toAttribute()));
+                        }
+                    }
+                    return new Eval(
+                        Source.EMPTY,
+                        new Aggregate(
+                            sampledAggregate.source(),
+                            sampledAggregate.child(),
+                            sampledAggregate.groupings(),
+                            sampledAggregate.originalAggregates()
+                        ),
+                        fields
+                    );
+                });
+                child = new PruneColumns().apply(child);
+            }
+
+            children.set(branchIndex, child);
+            return fork.replaceSubPlans(children).refreshOutput();
+        });
         logicalPlan.setOptimized();
         return logicalPlan;
     }
