+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/optimizer/rules/logical/SubstituteApproximationPlan.java
@@ -7,8 +7,8 @@
 
 package org.elasticsearch.xpack.esql.optimizer.rules.logical;
 
-import org.elasticsearch.xpack.esql.approximation.Approximation;
 import org.elasticsearch.xpack.esql.approximation.ApproximationPlan;
+import org.elasticsearch.xpack.esql.approximation.ApproximationVerifier;
 import org.elasticsearch.xpack.esql.optimizer.LogicalOptimizerContext;
 import org.elasticsearch.xpack.esql.plan.logical.LogicalPlan;
 import org.elasticsearch.xpack.esql.rule.ParameterizedRule;
@@ -24,11 +24,11 @@ public final class SubstituteApproximationPlan extends ParameterizedRule<Logical
         if (context.configuration().approximationSettings() == null) {
             // Approximation is not enabled
             return logicalPlan;
-        } else if (Approximation.verifyPlan(logicalPlan) == null) {
+        } else if (ApproximationVerifier.verifyPlan(logicalPlan) == null) {
             // Plan is not suitable for approximation
             return logicalPlan;
         } else {
-            // Returns an approximation plan with a placeholders for the sample probability.
+            // Returns an approximation plan with placeholders for the sample probabilities.
             // This placeholder will be replaced after executing the corresponding subplans.
             return ApproximationPlan.get(logicalPlan, context.configuration().approximationSettings());
         }
