+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/approximation/ApproximationPlanTests.java
@@ -18,6 +18,7 @@ import java.util.List;
 import java.util.Map;
 
 import static org.elasticsearch.xpack.esql.EsqlTestUtils.withDefaultLimitWarning;
+import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.nullValue;
@@ -38,7 +39,7 @@ public class ApproximationPlanTests extends ApproximationTestCase {
         assertThat(placeholder.nullable(), equalTo(Nullability.FALSE));
     }
 
-    public void testApproximationPlan_createsConfidenceInterval_withoutGrouping() throws Exception {
+    public void testApproximationPlan_createsConfidenceInterval_withoutGrouping() {
         LogicalPlan approximationPlan = ApproximationPlan.get(
             ApproximationTests.getLogicalPlan("FROM test | STATS COUNT(), SUM(emp_no)"),
             ApproximationSettings.DEFAULT
@@ -51,7 +52,7 @@ public class ApproximationPlanTests extends ApproximationTestCase {
         assertThat(approximationPlan, hasPlan(Eval.class, withField(("_approximation_certified(SUM(emp_no))"))));
     }
 
-    public void testApproximationPlan_createsConfidenceInterval_withGrouping() throws Exception {
+    public void testApproximationPlan_createsConfidenceInterval_withGrouping() {
         LogicalPlan approximationPlan = ApproximationPlan.get(
             ApproximationTests.getLogicalPlan("FROM test | STATS COUNT(), SUM(emp_no) BY emp_no"),
             ApproximationSettings.DEFAULT
@@ -64,7 +65,7 @@ public class ApproximationPlanTests extends ApproximationTestCase {
         assertThat(approximationPlan, hasPlan(Eval.class, withField(("_approximation_certified(SUM(emp_no))"))));
     }
 
-    public void testApproximationPlan_dependentConfidenceIntervals() throws Exception {
+    public void testApproximationPlan_dependentConfidenceIntervals() {
         LogicalPlan approximationPlan = ApproximationPlan.get(
             ApproximationTests.getLogicalPlan(
                 "FROM test | STATS x=SUM(emp_no) | EVAL a=x*x, b=7, c=TO_STRING(x), d=MV_APPEND(x, 1::LONG), e=a+POW(b, 2)"
@@ -87,7 +88,20 @@ public class ApproximationPlanTests extends ApproximationTestCase {
         assertThat(approximationPlan, hasPlan(Eval.class, withField(("_approximation_certified(e)"))));
     }
 
-    public void testColumnMetadata() throws Exception {
+    public void testApproximationPlan_withFork() {
+        LogicalPlan approximationPlan = ApproximationPlan.get(
+            ApproximationTests.getLogicalPlan(
+                "FROM test | FORK (STATS sum=SUM(emp_no)) (KEEP emp_no) (WHERE emp_no < 10 | STATS max=MAX(emp_no))"
+            ),
+            ApproximationSettings.DEFAULT
+        );
+        assertThat(
+            approximationPlan.output().stream().map(Attribute::name).toList(),
+            contains("sum", "_fork", "emp_no", "max", "_approximation_confidence_interval(sum)", "_approximation_certified(sum)")
+        );
+    }
+
+    public void testColumnMetadata() {
         LogicalPlan approximationPlan = ApproximationPlan.get(
             ApproximationTests.getLogicalPlan("FROM test | STATS count=COUNT(), sum=SUM(emp_no)"),
             ApproximationSettings.DEFAULT
